Rename MeasurementInputs2
